refactor: added home page files in a specific feature folder
