+++ src/features/store/homePage/components/wideCardRow/WideCard.tsx
@@ -1,9 +1,9 @@
 import Link from "next/link";
 import Image from "next/image";
 import { cn } from "@/shared/utils/styling";
-import { TWideCard } from "@/features/store/homePage/types/wideCard";
+import { TWideCard } from "../../types";
 
-const WideAd = ({ imgUrl, linkText = "Show Deals", smallTitle, title, url, isLightBG = false }: TWideCard) => {
+export const WideCard = ({ imgUrl, linkText = "Show Deals", smallTitle, title, url, isLightBG = false }: TWideCard) => {
   return (
     <div className={cn("relative h-[200px] flex-grow pl-7", isLightBG ? "text-gray-900" : "text-gray-100")}>
       <span className="relative z-5 text-sm mt-8 block">{smallTitle}</span>
@@ -21,5 +21,3 @@ const WideAd = ({ imgUrl, linkText = "Show Deals", smallTitle, title, url, isLig
     </div>
   );
 };
-
-export default WideAd;
