+++ src/features/store/homePage/components/wideCardRow/index.tsx
@@ -1,5 +1,5 @@
-import { TWideCard } from "@/features/store/homePage/types/wideCard";
-import WideAd from "../wideAd";
+import { TWideCard } from "../../types";
+import { WideCard } from "./WideCard";
 
 type TProps = {
   cards: TWideCard[];
@@ -10,7 +10,7 @@ export const WideCardRow = ({ cards }: TProps) => {
     <div className="w-full mt-15 flex flex-col gap-4 md:flex-row">
       {cards.map((card, idx) => {
         const { imgUrl, smallTitle, title, url } = card;
-        return <WideAd key={`${idx}-${title}`} imgUrl={imgUrl} smallTitle={smallTitle} title={title} url={url} />;
+        return <WideCard key={`${idx}-${title}`} imgUrl={imgUrl} smallTitle={smallTitle} title={title} url={url} />;
       })}
     </div>
   );
