+++ src/features/store/homePage/types/dealCards.ts
@@ -1,0 +1,6 @@
+import { TProductCard } from "@/types/common";
+
+export type TDealCard = TProductCard & {
+  dealDate: Date;
+  dealPrice: number;
+};
