+++ src/features/store/homePage/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./dealCards";
+export * from "./wideCard";
+export * from "./collection";
